+++ src/app/(dashboard)/instagram-feed/actions/update-feed.action.ts
@@ -6,30 +6,32 @@ import { z } from 'zod';
 import { InstagramFeedFormSchema, InstagramFeedFormImagesSchema } from '@/schemas/instagram-feed.schema';
 import type { InstagramFeedForm, InstagramFeedFormImages } from '@/schemas/instagram-feed.schema';
 
-interface FeedFormImagesInsert {
+interface UpdateInstagramFeedParams {
+  feedId: string;
+  formData: InstagramFeedForm;
+  formImages: InstagramFeedFormImages[];
+}
+
+interface FormImagesInsert {
   instagram_feed_id: string;
   media_id: string;
   position: number;
 }
 
-export async function updateInstagramFeed(
-  feedId: string,
-  feedFormData: InstagramFeedForm,
-  feedFormImages: InstagramFeedFormImages[],
-) {
+export async function updateInstagramFeed({ feedId, formData, formImages }: UpdateInstagramFeedParams) {
   if (!feedId) return { success: false };
 
-  const feedFormDataParsed = InstagramFeedFormSchema.safeParse(feedFormData);
-  const feedFormImagesParsed = z.array(InstagramFeedFormImagesSchema).safeParse(feedFormImages);
+  const formDataParsed = InstagramFeedFormSchema.safeParse(formData);
+  const formImagesParsed = z.array(InstagramFeedFormImagesSchema).safeParse(formImages);
 
-  if (!feedFormDataParsed.success || !feedFormImagesParsed.success) return { success: false };
+  if (!formDataParsed.success || !formImagesParsed.success) return { success: false };
 
   // const { error: feedError } = await supabase
   //   .from('instagram_feeds')
   //   .update({
-  //     name: feedFormData.name,
-  //     layout: feedFormData.layout,
-  //     button: feedFormData.button,
+  //     name: formData.name,
+  //     layout: formData.layout,
+  //     button: formData.button,
   //   })
   //   .eq('id', feedId);
 
@@ -42,7 +44,7 @@ export async function updateInstagramFeed(
 
   // if (mediaDeleteError) return { success: false };
 
-  // const instagramFeedMediaTableRows: FeedFormImagesInsert[] = feedFormImagesParsed.data.map((image) => ({
+  // const instagramFeedMediaTableRows: FormImagesInsert[] = formImagesParsed.data.map((image) => ({
   //   instagram_feed_id: feedId,
   //   media_id: image.media.id,
   //   position: image.position,