+++ src/app/(dashboard)/instagram-feed/hooks/use-feed-form..ts
@@ -2,19 +2,19 @@
 
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { z } from 'zod';
+import { useMutation } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
 import { createInstagramFeed } from '../actions/create-feed.action';
 import { updateInstagramFeed } from '../actions/update-feed.action';
 import { InstagramFeedFormSchema } from '@/schemas/instagram-feed.schema';
-import type { InstagramFeed } from '@/schemas/instagram-feed.schema';
+import type { InstagramFeed, InstagramFeedForm } from '@/schemas/instagram-feed.schema';
 import { useInstagramFeedStore } from '@/store/instagram-feed-store';
 
 export default function useInstagramFeedForm(type: 'create' | 'edit', formData?: InstagramFeed | null) {
   const { selectedImages } = useInstagramFeedStore();
 
-  const form = useForm<z.infer<typeof InstagramFeedFormSchema>>({
+  const form = useForm<InstagramFeedForm>({
     defaultValues: {
       name: formData?.name ?? 'New feed',
       layout: {
@@ -36,33 +36,43 @@ export default function useInstagramFeedForm(type: 'create' | 'edit', formData?:
     resolver: zodResolver(InstagramFeedFormSchema),
   });
 
-  const onSubmit = async (data: z.infer<typeof InstagramFeedFormSchema>) => {
+  const instagramFeedFormCreateMutation = useMutation({
+    mutationFn: createInstagramFeed,
+    onSuccess: (result) => {
+      if (result.success) {
+        toast.success('Successfully created');
+      } else {
+        toast.error('Error creating feed');
+      }
+    },
+  });
+
+  const instagramFeedFormUpdateMutation = useMutation({
+    mutationFn: updateInstagramFeed,
+    onSuccess: (result) => {
+      if (result.success) {
+        toast.success('Successfully updated');
+      } else {
+        toast.error('Error updating feed');
+      }
+    },
+  });
+
+  const onSubmit = async (data: InstagramFeedForm) => {
     if (selectedImages.length === 0) {
       toast.error('No feed images selected');
       return;
     }
 
     if (type === 'create') {
-      const res = await createInstagramFeed(data, selectedImages);
-
-      if (res.success) {
-        toast.success('Successfully created');
-      } else {
-        toast.error('Error creating feed');
-      }
+      instagramFeedFormCreateMutation.mutate({ formData: data, formImages: selectedImages });
     } else {
       if (!formData) {
         toast.error('Error updating feed');
         return;
       }
 
-      const res = await updateInstagramFeed(formData.id, data, selectedImages);
-
-      if (res.success) {
-        toast.success('Successfully updated');
-      } else {
-        toast.error('Error updating feed');
-      }
+      instagramFeedFormUpdateMutation.mutate({ feedId: formData.id, formData: data, formImages: selectedImages });
     }
   };